f608飛黃騰達 完成了

+++ vs-code/g277笛卡爾樹.cpp
@@ -154,7 +154,7 @@ public:
     {
         dfs_Sum(top->l_Node);
     }
-    int dfs_Sum(node* now)
+    long long dfs_Sum(node* now)
     {
         if(now)
         {
